app-tier/requirements.txt
tqdm==4.59.0
#pdftotext==2.1.5
requests==2.25.1
urllib3==1.26.4
beautifulsoup4==4.9.3
protobuf==3.15.8
google-cloud-storage==1.37.1
google-cloud-pubsub==2.4.1
google-cloud-datastore==2.1.0
#pdftotext==2.1.5